requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
pyarrow==14.0.1
requests==2.31.0
python-multipart==0.0.6
anthropic==0.3.11
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openpyxl==3.1.2
reportlab==4.0.7
python-telegram-bot==20.7
schedule==1.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
apify-client==1.7.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
openai==1.3.5
chromadb==0.4.18
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.25.2
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
jwt==1.3.1
plotly==5.17.0
monitoring